requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
attrs==19.3.0
autograd==1.3
autograd-gamma==0.4.1
backcall==0.1.0
bleach==3.1.0
Bottleneck==1.2.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.7.0
coloredlogs==10.0
conda-package-handling
cryptography==2.4.2
h5py==2.9.0
humanfriendly==4.18
idna==2.8
image==1.5.27
imagecodecs-lite
imageio==2.6.1
importlib-metadata==1.3.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
joblib==0.14.1
json5==0.8.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lifelines==0.22.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.2
munch==2.5.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.3.1
patsy
pretrainedmodels==0.7.4
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.6.8
python-dateutil==2.7.5
pytz
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.6.0
query-string
requests
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
solt==0.1.5
spams==2.6.1
statsmodels==0.10.1
tensorboard==2.1.0
tensorboardX==2.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.2
tifffile==2019.7.26.2
toolz==0.10.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
typing==3.6.4
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
